phpbb2privatemsg.module - import unimported users' messages as anonymous user.
Throwing away messages from unimported users can cause errors like this:

    PDOException: SQLSTATE[HY000]: General error:
    1366 Incorrect integer value: '' for column 'thread_id' at row 1:
    INSERT INTO {phpbb2privatemsg} (mid, thread_id, msg_id)
    VALUES (:db_insert_placeholder_0,
            :db_insert_placeholder_1,
            :db_insert_placeholder_2);
    Array ( [:db_insert_placeholder_0] => 3
            [:db_insert_placeholder_1] =>
            [:db_insert_placeholder_2] => 4 )
    in phpbb2privatemsg_import()
    (line 154 of .../drupal-7.8/modules/phpbb2drupal/phpbb2privatemsg.module).

To elaborate, the thread_id is set from a look-up in the
phpbb2privatemsg table which records the thread-id, phpbb 'msg_id',
and drupal 'mid' message IDs.  However, if a message was skipped
because its drupal user id cannot be found, this look-up will fail to
retun anything, and the folloing attempt to set the thread_id field of
a record for the last imported message in the same table to an
undefined value throws an exception.

These users may not exist because they have not posted.  Or perhaps
some other reason, maybe a bug elsewhere.  I am not currently sure if
the check which decides which users have posted deems they have if
they have sent private messages (it should do), but in any case I
think we should not just fall over here, even if this is really a
problem elsewhere.

What to do?  My best suggestion is to import all messages, but to
import unknown user ids as the 'anonymous' user, or perhaps some other
dummy user specially created for the purpose.

This patch implements this. Currently it is hardwired to import them
as the anonymous user.  Ideally the user id would be configurable.

(After importing, perhaps there could be a sweep to delete threads with
no known users within them. )